meeting_copilot/requirements.txt
whisperx>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
crewai>=0.11.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
soundfile>=0.12.1
ffmpeg-python>=0.2.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0
chromadb>=0.4.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1 